Reduce to only tRNA code
The code is now reduced to 1017 lines. Next I want to functionalize all
the stem/loop prediction code. These should become simple functions of
sequence that return lists of matches and energies. This should work
fine for T-stem, C-stem, and D-stem. The A-stem is more complicated
since it is a pseudoknot of unknown length. The V-loop is similarly
complicated.