Begin adding support for tmRNA and mtRNA
The tRNA, mtRNA and tmRNA algorithms all have little overlap in the
source code. There is little, if any, performance benefit to threading
these searches together. Further, at least in the case of mtRNA, there
are few good reasons to search a single sequence for both mtRNA and
tRNA. The sequence either is a mitochondria or it is not. Searching
entire nuclear genomes for mtRNA patterns will most likely find many
non-specific matches (the pattern is very tenuous). So I will write each
of the three functions completely independently. Also, I will place each
in their own C file.

In this commit, I add only mtRNA. The return value of tRNA and mtRNA are
the same. A transfer RNA with an anticodon in the C-loop. The mtRNA may
have 0-length stems on either the T or D structure, but these structurs
can still be described by the tRNA class.

tmRNA, however, will need a new structure. I should store the start and
length of the C-loop insert and also the location and length of the tag.
I will work on this after I add mtRNA.