Resolved tmRNA segfaults
It currently cannot identify the tmRNAs in the tmRNA dataset, probably
because the offsets are too large and the boundary checks break out to
strictly. I probably need a judicious addition of limiting search
distance to the size of the sequence.

I can predict one tmRNA in the Mycoplasmoides pneumoniae genome and this
prediction matches the original. There is an off-by-one potential error
in the start position that I should investigate. Also the length is
listed as -569164.

Next up I need to fix the boundary problems that prevent testing of my
large tmRNA set of genes. Then I need to ensure that my results match
the original program results.

This commit also adds many commented-out print statements. I will leave
this in for now, but they will need to be removed once I work out the
remaining kinks.